+++ web-react-ts/src/pages/directory/directoryRoutes.ts
@@ -142,6 +142,9 @@ const DisplayAllStreamMinistries = lazy(
 const DisplayAllCouncilHubs = lazy(
   () => import('pages/directory/display/AllCouncilHubCouncils')
 )
+const DisplayAllConstituencyHubs = lazy(
+  () => import('pages/directory/display/AllConstituencyHubs')
+)
 const DisplayAllMinistries = lazy(
   () => import('pages/directory/display/AllMinistries')
 )
@@ -631,6 +634,12 @@ export const directory: LazyRouteTypes[] = [
     path: '/hubcouncil/displayall',
     element: DisplayAllHubCouncils,
     roles: [...permitMe('HubCouncil'), ...permitMe('Council')],
+    placeholder: false,
+  },
+  {
+    path: '/constituency/hubs',
+    element: DisplayAllConstituencyHubs,
+    roles: [...permitMe('Constituency'), ...permitMe('Hub')],
     placeholder: false,
   },
   {